Add historical constituency/population data
This takes the last set of data and projects it back which isn't correct
but the exercise of getting correct population figures for each parliament
is left for a future exercise in trawling the ONS website.

We have to reorder the loading of data since we need to know what period
we're loading the data from so the petition data is loaded first now.